Remove v1alpha CR logic
